Fix item stability during metadata updates
